ProjElearning/requirements.txt
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
